JCF: improvement to transfer plugin memory bookkeeping when advanced memory usage is false
Prior to this commit, the size of a transfer plugin's buffer as
defined by max_fragment_size_words was *always* just the
max_fragment_size_bytes value set in the DAQInterface settings file
(modulo the byte-word conversion). While this made sense for the
transfer plugins between boardreaders and eventbuilders, it didn't
make sense for the transfer plugins designed to handle entire events,
e.g., the transfer plugins between EventBuilders and DataLoggers. For
this reason, these "event transferring" transfer plugins have their
memory buffer set to the value defined by max_event_size_bytes in the
FHiCL (if available), and if not, to simply the
max_fragment_size_bytes value from the settings file times the number of
boardreaders.

Note that there's a good argument to be made that if
max_event_size_bytes is set in an EventBuilder, DataLogger or
Dispatcher's FHiCL that it should perhaps take precedence over the
setting in the DAQInterface file. As of this commit, this isn't the
case; however, this commit is a definite improvement in the situation
when non-advanced memory usage is selected.