JCF: added subsystem and rank to the Procinfo structure
Laying the groundwork for proper bookkeeping of artdaq processes
between subsystems, I've added subsystem and rank to the Procinfo
structure. It's now the case that the file referred to by
$DAQINTERFACE_KNOWN_BOARDREADERS_LIST (see
docs/known_boardreaders_list_example) requires a fourth column,
representing the subsystem number of the component in question.

Modulo the above change, I've confirmed that my changes are backwards
compatible (cf. runs 1540 and 1542 in mu2edaq01:/home/jcfree/run_records)