JCF: update bookkeeping to accommodate simple 2-subsystem arrangements
Since in a commit from earlier today both rank # and subsystem # were
added to the artdaq process structure Procinfo, I've rewritten the
bookkeeping functionality to take advantage of this so that:

1) artdaq process types within a subsystem bear the same relationship
to each other as before (e.g., subsystem 2 dataloggers appear in the
destination table for subsystem 2 eventbuilders)

2) An eventbuilder in subsystem 1 adds eventbuilders in subsystem 2 to
its destination table, on top of any subsystem 1 dataloggers

3) The routingmaster mediates between subsystem 1 boardreaders and
subsystem 1 eventbuilders

I've tested that the sources and destinations tables are constructed
correctly (see, e.g., mu2edaq01:/home/jcfree/run_records/1563 ), and
also that the new algorithm is backwards compatible - i.e., if all
processes are subsystem 1, then you recover the previous connections
between the processes.

The next logical step in the development of subsystem support in
DAQInterface, of course, is to generalize this for multiple subsystems.