JCF: decouple the creation of the ranks.txt file from pmt.rb's process management
With ranks now available in the process data structures as of the last
week or two, there's no longer any point in deriving the process ranks
from the pmtConfig.XXXXX file since they can be taken directly from
self.procinfos. This has a dual advantage: (1) pmt.rb won't be the
only way to launch processes in DAQInterface moving forward, so we
can't rely on it to give us ranks, and (2) we can have a ranks file
even in the absence of any process management at all.