JCF: in direct process management, when a process dies, thoroughly clean up after it
Before this commit, what I was observing was that while runs didn't
end after a process was killed - which is good, as this is the desired
behavior - it was also the case that often I was running into
difficulty starting a new run. This is because some resources weren't
released - leftover art processes associated with the now-deceased
artdaq process, for example. With this commit, I clean up more
thoroughly: in practice, this means killing off every process which
uses the relevant XML-RPC port as revealed by "netstat -alpn", as well
as the ssh in the case of processes on remote hosts.