JCF: refuse to try launching processes if one or more processes are already using the needed ports
One potential source of drama in direct process management is if
DAQInterface tries to launch an artdaq process on a given host,
listening on a given port...and due to improper cleanup, a zombie
process is already on that socket. With this commit, DAQInterface will
check for this phenomenon, and if it finds pre-existing processes it
will throw an exception rather than initiate a doomed-to-fail process
launch.