JCF: in case of rare double-process-launch, print warning rather than assert false
In a hard-to-reliably-reproduce bug, it appears that occasionally
DAQInterface will try to launch a set of artdaq processes on a given
host not once but twice when we're launching processes directly rather
than via pmt.rb. Rather than asserting false when this happens, just
print a warning so I can better study what's going on.