JCF: support per-partition messageviewers to satisfy Issue #21360
Before this commit, it was always the case that port 30000 was used to
send MessageFacility messages via UDP to a messageviewer
instance. This is obviously incompatible with multiple partitions
operating in parallel on a host, so now each partition gets its own
messageviewer listening on its own port, where as per
https://cdcvs.fnal.gov/redmine/projects/artdaq/wiki/Partition_notes_1
the port is 10005 + (partition number)*1000.