JCF: provide more helpful/informative error messages during booting
I've modified DAQInterface so that during a boot failure it tells you
specifically which processes didn't launch, and also how to recreate
what DAQInterface did.

Note that this is currently only supported for
DAQINTERFACE_PROCESS_MANAGEMENT_METHOD=pmt, thought getting it to work
for the direct case is near the top of my priority list.