JCF: Separate the creation of the log directories from the launching of the processes
Previously, when the only method of launching processes was via
pmt.rb, the making of the needed log directories (for run-specific
softlinks, etc.) was part of the launch function. However, now that
more than one method is possible, it's pointless to have to
cut-and-paste this needed functionality each time a new launch method
is introduced, so this log directory making functionality has been
moved into the body of the boot transition function, preceding the
launch of the processes.