JCF: add support for running on large numbers (hundreds) of ToySimulators
Specifically, what I've done is altered the component_push and
component_pull subconfigurations so that each supports up to 240
ToySimulators, numbered 01-240 and 1001-1240, respectively. I've also
added a known boardreaders list example in the docs/ directory which
supports usage of these ToySimulators on the mu2edaq cluster
(known_boardreaders_list_mu2edaq). This is all going to be used to
test how DAQInterface responds to DUNElike numbers of artdaq
processes.