JCF: in direct process management, have the heartbeat function check ALL processes on a node with a single ssh call
An inefficiency that became glaringly obvious today as I ran 160
boardreaders across the mu2edaq cluster (see, e.g., run 1825) is that
the heartbeat function for direct process management was very slow to
return as it performed a separate ssh call for each process checking
to see whether it existed. With this commit, the heartbeat function
will check ALL the artdaq processes on a given node via a single ssh
call.

Please note that a similar improvement in logic can be applied to the
way processes get killed...