JCF: add more info for the user on what DAQInterface is doing
Now that we've started testing runs with large numbers (hundreds) of
artdaq processes, the importance of benchmarking DAQInterface's
behavior is heightened. I've started adding more printouts concerning
where DAQInterface is during its various transitions (e.g., if it's
testing that the setup script sources correctly on all nodes during
boot, or if it's performing bookkeeping during configure) so we can
(A) be confident that it's doing something, since some of these
actions take a while with large numbers of processes, and (B)
determine where DAQInterface is spending its time, and perhaps
optimize accordingly.