JCF: to save time, at the start of boot source the setup script on just one node
In the fall of 2017, I implemented a feature in DAQInterface where it
would source the setup script on every node on which artdaq processes
were intended to run. While useful in the sense that it could catch
whether a setup script was returning a nonzero value, it becomes time
consuming when lots of nodes are in use (about 24 seconds, e.g., when
using all available nodes on the mu2edaq cluster). As a compromise,
DAQInterface will now instead just source on one node to make sure
things are OK. Note that this means that if TRACE functionality has
just been introduced to the setup file the traces *may* not show up in
the first run taken after the modification; I'm not certain whether
this is the case, however.