JCF: make sure that if something goes wrong when trying to make the log directories that details are printed.
If there's a "No route to host" ssh error with one of the nodes, the
first place it'll likely appear (assuming the node in question wasn't
the one selected for the setup sourceme test) will be when
DAQInterface tries to make the logfile directories on the
node. However, before this commit, the user was simply told that there
was a problem running "mkdir -p", which can be misleading if the
problem is with the ssh call itself.