JCF: (begin) removing unnecessary X11 forwarding in ssh calls
This change is motivated by the fact that I was finding that during
the recover "transition", if I was using O(150) processes, recovery
took a few minutes, with many, many messages of the sort
"/usr/bin/xauth: timeout in locking authority file
/home/jcfree/.Xauthority" and "Warning: No xauth data; using fake
authentication data for X11 forwarding.". I can trace these to the use
of ssh to (A) determine whether artdaq processes still existed, and
(B) to kill the existing processes. For neither of these actions is it
necessary or useful to allow X11 forwarding, so I've turned it off,
and now in comparable situations recovery takes maybe 10-20 seconds
rather than minutes.