JCF: Add the process_launch_diagnostics_base function, to provide detail on what went wrong during a process launch
As described in Issue #21739, when using direct process management, in
the event that not all processes successfully launched potentially
useful info was getting suppressed. I've now added a function required
by process management modules called
"process_launch_diagnostics_base", which is called when not all
processes launch and is meant to provide the user with further info on
what went wrong. For direct process management, this means that the
stdout and stderr output of the attempted process launch on every node
where something went wrong gets printed; in the case of pmt (for the
time being at least) this is just a no-op.