JCF: when using direct process management, save output in a file called /tmp/launch_attempt_<username>_partition<partition number>
This file is saved on each node, and includes the output of the source
of the setup script and the output of the artdaq processes. Note that
when I say "output", I'm including stderr and stdout, not just
MessageFacility messages. If something goes wrong during process
launch, users are directed to look at the file. Note that for each new
launch attempt it gets clobbered.