JCF: if a subsystem has eventbuilders and dispatchers but no dataloggers, connect the eventbuilder to the dispatcher
This is in respond to Issue #21856, where Eric noticed that runs with
the eventbuilder_diskwriting configuration had problems in that events
weren't making it from the eventbuilder to the dispatcher (no
dataloggers exist in the run). With this commit, I've added logic to
bookkeeping whereby, if DAQInterface determines there are no
dataloggers in a subsystem in question, then if there are dispatchers
they'll expect to receive events from eventbuilders rather than
dataloggers.