JCF: cleaner, more informative error reporting
In a nutshell, errors which occur during transitions were creating
stack traces which probably hurt more than they helped in that they
were (A) messy, and (B) traced through DAQInterface function calls in
a way that didn't reveal the root of the problem, typically
originating in the artdaq process. Instead of printing the stack
traces, before DAQInterface even throws the exception it will try to
determine if the problem is because a process died, or an exception
was thrown within the process, etc., also pointing users to logfiles
where appropriate. Note that if a debug level of 2 or higher is
selected, the stack traces will reappear.