JCF: add default rules for which process deaths should (or should not) end a run
Basically, if a process dies and its label doesn't match a regexp in
the file $DAQINTERFACE_PROCESS_REQUIREMENTS_LIST (or this file doesn't
exist), then DAQInterface will determine whether or not to end the run
based on the type of process which died. The rules are simple:

-If a process is a BoardReader, then end the run
-If a process writes data to disk, whether EventBuilder or DataLogger, end the run
-Any other process dies, continue running

If a process dies and its label matches at least one regexp in
$DAQINTERFACE_PROCESS_REQUIREMENTS_LIST, the rule(s) which apply to it
will completely override this default behavior.