JCF: check for a file containing commit info when saving the run record
This is motivated by the fact that in moving to a system where builds are
performed on local disks rather than on NFS at ProtoDUNE's np04 cluster, it's
no longer the case that DAQInterface can safely assume that it can find a
package's git repo in the ./srcs subdirectory of the directory containing the
setup script. If it can't, it'll check to see if there's a file of the form
<pkgname>_commit_info.txt in the directory instead. Note that in the last day or
so I've added commits to dune-artdaq's develop branch and dune-raw-data's
for_dune-artdaq branch s.t. whenever you build them they'll create files with just
these names. I've tested the code in this commit (though they included a couple
of print statements I've since removed) using runs 8678 and 8679 on the ProtoDUNE
cluster.