JCF: on config, add timestamped copy of the run record, which only gets deleted when the run starts and the run record is permanently saved
If, for whatever reason, an attempt at a run never makes it into the
running state (e.g., a config fails), it's useful to have the info in
the run record available (exact FHiCL used, setup script, names of
logfiles, etc.). The run record gets permanently saved in the run
records directory, but that only happens when we have a run number,
i.e., we make it into the running state. Otherwise, it's in
/tmp/run_record_attempted_\$USER. Before this commit, this record
would get clobbered every time there was a new config; now it's the
case that we save the run record with a unique timestamp, and it only
gets deleted IF we enter the running state. Otherwise, if we don't
enter the running state, we possess it indefinitely for diagnostic
purposes.