JCF: get the table_range function to make sure it's found a table, and not just a token with the right name
This change is prompted by an email thread between myself and Kurt on
June 26, where we were running into a bug in which the table_range
function, seeking out a table named "art" in the pdune_swtrig_noRM
config's DFO.fcl, was actually stopping on a key: value pair
("use_art: true") instead of the desired table farther along in the
FHiCL document. While the issue was sidestepped by using a relevant
nested table ("outputs") as the argument to table_range instead of
"art", Kurt made the point that it would be ideal to have table_range
actually look for a table and not just any matching token. I've made
that change here, and confirmed that we get the same FHiCL code from
bookkeeping searching for table "art" with this commit's new
table_range function as we did with the old table_range function
searching for "outputs". This comparison was done by performing runs
with the pdune_swtrig_noRM config, specifically, one can compare
/home/jcfree/run_records/2714 and /home/jcfree/run_records/2715 on the
mu2edaq cluster.