JCF: make sure allowed lack of ups doesn't result in unwanted stderr output
In get_package_version, an exception was getting raised because the
bash_unsetup_command string was yielding output to stderr thanks to
the "which ups" at its beginning. However, ups doesn't *have* to be
setup, so it was an unwanted exception. What I've done is ensure that
any output to stderr from "which ups" gets swallowed, so DAQInterface
doesn't get needlessly upset.