JCF: lightly-tested (i.e., don't use it yet) bookkeeping which prioritizes private network use
With this commit, DAQInterface will try to get routing to happen over
a private network if available. This is described in more detail in
the comment from Aug. 8, 5:29 PM in Issue #22226. I successfully
performed a run where boardreaders, eventbuilders and a routing_master
were used, all on the same subsystem
(sbnd-daq33.fnal.gov:/home/nfs/jcfree/run_records/25) using this
commit's code. More testing needs to be done, and in particular, from
the perspective of Issue #22226 I have yet to deal with the
multicast_interface_ip used for requests, or to support the convention
whereby a parameter is only bookkept if its value is set to
"BOOKKEPT_BY_DAQINTERFACE"