JCF: update bookkeeping of the private networks to account for last subnet
There are actually two changes in this commit. The first is, when
trying to figure out whether all processes in a routing master or
request group can see the same private network (see
https://cdcvs.fnal.gov/redmine/issues/22226 for more detailed
defitions of what these are), the last subnet is now ignored - i.e., a
bug's been removed. The second is, it appears that for the
routing_master_hostname for routing_master, you need to NOT ignore the
last subnet - I didn't realize that on the previous commit.