JCF: provide individual artdaq process states on demand
When status_with_side_effects.sh is executed (or, more generally, the
state_with_side_effects XML-RPC command is sent to DAQInterface - yes,
slightly different name than the script) DAQInterface will do the following:

It produces an ASCII file of the form /tmp/artdaq_process_states_<username>_partition<partition #>,
with a list of the existing processes and their statuses. If "debug_level" is set to 5 or higher, the
list will also appear in DAQInterface's stdout - note this very high level since the output tends to
interfere with output from DAQInterface's main thread of execution.

This commit is designed to satisfy Issue #23146