JCF: allow DAQInterface to continue working with versions of artdaq_database older than v1_04_75
I recently added usage of the conftool module's getListOfAvailableRunConfigurationsSubtractMasked function
in DAQInterface; however, since that function was only introduced in v1_04_75, it means using an older
database version was resulting in an immediate DAQInterface failure on launch. With this commit, it'll
simply print a warning informing users that it'll ignore whether or not a configuration's been masked off, and
then proceed.