JCF: make sure that the DAQInterface run start and run stop times are saved in the database
Gennadiy pointed out to me about a week ago that the DAQInterface start time and DAQInterface stop
times which get saved in the run record's metadata.txt file don't get saved into the database. With
this commit, these values are saved in the RunHistory2.fcl file. Prior to his, RunHistory2.fcl was
produced only when DAQINTERFACE_PROCESS_MANAGEMENT_METHOD == "external_run_control", and it contained
the FHiCL-ized content of /tmp/info_to_archive_partition<partition number>.txt file, if it existed. Now,
RunHistory2.fcl is produced for all DAQINTERFACE_PROCESS_MANAGEMENT_METHOD values, where in the case of
"pmt" and "direct" it exclusively contains the start/stop times, and in the case of "external_run_control"
these times are prepended to the "traditional" contents of RunHistory2.fcl