JCF: make sure a requested transition doesn't proceed after we've finished recovery
The use case here is that a transition request gets sent to
DAQInterface, DAQInterface sets a boolean indicating that on the next
pass through the runner function it should execute that transition,
and then at some point before that next pass something goes wrong and
a recovery occurs, putting DAQInterface in the Stopped state. You
obviously DON'T want the requested transition to go through at that
point- so when we enter the recover function, the booleans that govern
what transition will execute on the next pass through the runner
function all get set to False.