JCF: Issue #23147, improvements to just_repeat_run.sh
The main complaint in the issue, that just_repeat_run.sh didn't care
if DAQInterface is being run on a host different than the one it ran
on during the run-to-be-repeated, is addressed in that this difference
is treated the same as a code difference or a known boardreaders list
difference: it won't run unless --nostrict is selected.

However, I've improved the output in that now, even if --nostrict HAS
been selected, all deviations are printed out. This prevents a
scenario in which someone selects --nostrict because they don't care
about one particular deviation, and then other deviations they may
care about get suppressed.