JCF: Issue #23784: if swig_artdaq is imported, ARTDAQ_LOG_FHICL needs to be properly set
In order for an instance of swig_artdaq to send messages to
messageviewer, ARTDAQ_LOG_FHICL needs to be set to point to a FHiCL
document which defines the udp port MessageViewer's listening
on. Consequently, if swig_artdaq is imported, DAQInterface will create
this document if necessary, and also require that ARTDAQ_LOG_FHICL
already be pointing to it.